Final case sprites, Case price Progression, Balance start

+++ YandexGames/Assets/_Scripts/ScriptableObjects/MoneyCaseData.cs
@@ -5,7 +5,7 @@ namespace _Scripts.ScriptableObjects
     [CreateAssetMenu(fileName = "NewMoneyCase", menuName = "ScriptableObjects/Money Case")]
     public class MoneyCaseData : CaseData
     {
-        public int Price;
+        public float Price;
         public bool IsTotalCases;
     }
 }
